backend/DataMiningWebMining/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
PyGithub==1.59.1
networkx==3.2.1
python-louvain==0.16
matplotlib==3.8.2
pyvis==0.3.2
